issue #58 chore : plan 구조 일부 변경
