+++ server/src/routers/index.js
@@ -1,11 +1,6 @@
 const router = require('express').Router()
+
 const user = require('./user')
-/**
- * @swagger
- * tags:
- *   name: Users
- *   description: 유저 추가 수정 삭제 조회
- */
 router.use('/user', user)   
 
 module.exports = router
