+++ server/src/routers/user/user.model.js
@@ -2,8 +2,26 @@
 
 const { QueryTypes } = require("sequelize");
 const db = require("../../util/database");
-const raw2str = require("../../util/rawtostr");
+const { raw2str } = require("../../util/rawtostr");
 const token = require("../../util/jwt");
+
+/**
+ * @typedef {Object} User
+ * @public
+ * @property {string} UserId - 사용자의 고유 ID
+ * @property {integer} AccountType - 사용자의 계정 타입
+ * @property {string} OauthId - 소셜 로그인일 경우 소셜 로그인 고유 ID
+ * @property {string} Username - 일반 로그인일 경우 사용자의 이메일(아이디)
+ * @property {string} Password - 일반 로그인일 경우 사용자의 비밀번호
+ * @property {string} PhoneNum - 사용자의 전화번호
+ * @property {string} Nickname - 사용자의 닉네임
+ * @property {string} ProfileImage - 사용자의 프로필 이미지
+ * 
+ * @private
+ * @property {Object} dbsetup - sequelize DB 연결 객체
+ * 
+ * 이 클래스는 User 테이블에 대한 모델입니다.
+*/
 class User {
     constructor(param) { //Object.assign(this, param);
         this.UserId = param.UserId;
@@ -21,8 +39,7 @@ class User {
         try {
             /* 일반 로그인 */
             if (this.AccountType == 0) {
-                let data = await db.query(`SELECT * FROM DB_USER WHERE USERNAME = '${this.Username}' and PASSWORD = '${this.Password}'`, { type: QueryTypes.SELECT });
-                data = await raw2str.sqlswap(data);
+                let data = raw2str(await db.query(`SELECT * FROM DB_USER WHERE USERNAME = '${this.Username}' and PASSWORD = '${this.Password}'`, { type: QueryTypes.SELECT }));
                 if (data.length == 0) {
                     return { status: false, reason: "아이디나 비밀번호가 맞지 않습니다." }
                 }
@@ -38,29 +55,29 @@ class User {
             }
             /* 카카오,네이버 로그인 */
             else if (this.AccountType == 1 || this.AccountType == 2) {
-                let query = await raw2str.sqlswap(await db.query(`SELECT * FROM DB_USER WHERE OAUTH_ID = '${this.OauthId}'`, { type: QueryTypes.SELECT }));
+                let query = raw2str(await db.query(`SELECT * FROM DB_USER WHERE OAUTH_ID = '${this.OauthId}'`, { type: QueryTypes.SELECT }));
                 if (query.length == 0) {
                     /* 계정이 없을 경우 회원가입 */
                     query = await db.query(`INSERT INTO DB_USER (ACCOUNTTYPE, OAUTH_ID, USERNAME, PASSWORD, PHONENUM, NICKNAME) VALUES ('${this.AccountType}', '${this.OauthId}', '', '', '', '')`, { type: QueryTypes.INSERT });
-                    query = await raw2str.sqlswap(await db.query(`SELECT * FROM DB_USER WHERE OAUTH_ID = '${this.OauthId}'`, { type: QueryTypes.SELECT }));
+                    query = raw2str(await db.query(`SELECT * FROM DB_USER WHERE OAUTH_ID = '${this.OauthId}'`, { type: QueryTypes.SELECT }));
                     if (query.length == 0) {
                         return { status: false, reason: "계정 등록에 실패했습니다." };
                     }
                     let isProfileExists = true;
-                    if (data[0].NICKNAME == null || data[0].NICKNAME == "") {
+                    if (query[0].NICKNAME == null || query[0].NICKNAME == "") {
                         isProfileExists = false;
                     }
-                    let AccessToken = token.generateAccessToken(q3[0].USER_ID);
-                    let RefreshToken = token.generateRefreshToken(q3[0].USER_ID);
+                    let AccessToken = token.generateAccessToken(query[0].USER_ID);
+                    let RefreshToken = token.generateRefreshToken(query[0].USER_ID);
                     return { status: true, reason: "로그인 성공", AccessToken: AccessToken, RefreshToken: RefreshToken, isProfileExists: isProfileExists };
                 }
                 else {
                     let isProfileExists = true;
-                    if (data[0].NICKNAME == null || data[0].NICKNAME == "") {
+                    if (query[0].NICKNAME == null || query[0].NICKNAME == "") {
                         isProfileExists = false;
                     }
-                    let AccessToken = token.generateAccessToken(q1[0].USER_ID);
-                    let RefreshToken = token.generateRefreshToken(q1[0].USER_ID);
+                    let AccessToken = token.generateAccessToken(query[0].USER_ID);
+                    let RefreshToken = token.generateRefreshToken(query[0].USER_ID);
                     return { status: true, reason: "로그인 성공", AccessToken: AccessToken, RefreshToken: RefreshToken, isProfileExists: isProfileExists };
                 }
 
@@ -133,7 +150,7 @@ class User {
     }
     async getInfo() {
         try {
-            let data = await raw2str.sqlswap(await db.query(`SELECT * FROM DB_USER WHERE USER_ID = '${this.UserId}'`, { type: QueryTypes.SELECT }));
+            let data = await raw2str(await db.query(`SELECT * FROM DB_USER WHERE USER_ID = '${this.UserId}'`, { type: QueryTypes.SELECT }));
             return (data.length == 0) ?
                 { status: false, reason: "유저 아이디를 찾을 수 없습니다." } :
                 { status: true, reason: "유저 정보 조회 성공", result: data[0] };
