--- server/src/routes.js
@@ -1,15 +1,0 @@
-/*
-
-// 계획 관련 API
-const plan = require("./api/plans");
-router.put("/api/plans", plan.create);
-router.delete("/api/plans", plan.delete);
-router.post("/api/plans", plan.update);
-router.get("/api/plans", plan.getPlan);
-router.get("/api/plans/list", plan.getPlanList);
-router.put("/api/plans/participants", plan.addParticipant);
-router.delete("/api/plans/partici*pants", plan.deleteParticipant);
-
-
-module.exports = router;
-*/
